Make std::set_unexpected and std::set_terminate set global (process) unexpected and terminate handlers, respectively.  Added pathscale:: versions of both (not exposed in the headers, so not - yet - part of the public API), which retain the old behaviour and set a thread-local version.
Also added pathscale::set_use_thread_local_handlers(), which allows the
user to request the thread-local behaviour from the std:: versions.
Also not exposed via a header yet.  The pathscale:: namespace might not
be the most sensible place for these, and some thought should be given
to this before actually making them public.