Fix the uncaught exception count with rethrowing.
The rethrow function (unlike the throw functions) wasn't incrementing
the uncaught exception count, which could cause the exception number to
become negative.  While investigating this, it transpired that the begin
catch function was unconditionally decrementing the number of uncaught
exceptions, even for foreign exceptions (which won't have incremented
the count in the first place, because they are not thrown using the C++
runtime).

The uncaught exception count is now correctly reported for C++
exceptions.  It would be nice if we could also report it for foreign
exceptions, but we can't force other languages to report that they've
caught an exception, so we can't trust our internal state machine to
give an authoritative report of whether a foreign exception has been
rethrown but not caught.

Fixes #45
Fixes #49