Fix unitialized variable in __cxa_demangle_gnu3 after #6
After #6, if the path "Try demangling as a type for short encodings" is
entered, and `cpp_demangle_read_type()` fails, `rtn` is uninitialized
but still returned after jumping to the `clean` label. The returned
garbage causes problems when `__cxa_demangle()` then supposes the result
is valid, and tries to `free()` it later.

On FreeBSD this showed up as a CI failure in googletest's gmock-matchers
test, which started aborting on the bad `free()`:

```
<jemalloc>: jemalloc_rtree.c:205: Failed assertion: "!dependent || leaf != NULL"
Process with PID 42015 exited with signal 6 and dumped core; attempting to gather stack trace
```

In this case, it is attempting to demangle `NSt3__15tupleIJibEEE`, and
apparenty this is a "naked" variant of `std::__1::tuple<int, bool>` so
it fails.