Add an option for disabling emergency buffers.
Section 3.3.1 of the Itanium ABI spec requires that we provide a small
set of emergency buffers to use when dynamic allocation of shared
objects fails.  We implement these because the spec says so but they
introduce a dependency on condition variables (which is not ideal on all
platforms) and they increase the binary size for negligible real-world
benefit (on systems with overcommit, allocation will succeed bug
accessing the page will fail and very few C++ codebases even attempt to
catch out-of-memory exceptions).

It is now possible to disable these buffers.  This option is off by
default, since the spec requires it.