Disable system output comparison test for FreeBSD CI
The test suite provides the option to compare test results against the
system C++ ABI implementation. On FreeBSD this comparison is of
questionable value, as the system C++ ABI implementation is libcxxrt.

Furthermore, this comparison can give false negatives due to a race in
test_guard.cc causing an off-by-one in number of tests run against the
system and libcxxrt implementations. Since the comparison is implemented
as a diff(1) of the output, this difference in the number of tests run
causes spurious failures.

Closes:		https://github.com/libcxxrt/libcxxrt/issues/42
Sponsored by:	The FreeBSD Foundation