Avoid noreturn warning on terminate()
In FreeBSD we have seen the following warning about the `terminate()`
function:

    contrib/libcxxrt/exception.cc:1538:2: warning: function declared 'noreturn' should not return [-Winvalid-noreturn]
     1538 |         }
          |         ^

Indeed the function is declared noreturn, but one of the paths through
the function falls off the end. Unfortunately the `terminatehandler`
type itself cannot be made noreturn (compilers do not accept this), but
the warning is easily fixed by moving the `abort()` call to end instead.

While here, apply the same logic for `unexpected()`, which is also not
supposed to return, and correct the comment there.